refactor remote control id to self contained component
