+++ src/components/remote-control-id/index.jsx
@@ -1,6 +1,8 @@
 import { useState, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 
+import './index.css';
+
 const Sides = {
     Left: 'Left',
     Right: 'Right',
